html/semantics/embedded-content/the-video-element/video-loading-lazy-autoplay-when-visible.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #target {
    position: absolute;
    top: 1000vh;
  }
</style>

<video id="target" src="/media/A4.mp4" loading="lazy" muted playsinline autoplay width="100" height="100"></video>

<script>
  const video = document.getElementById("target");

  const t = async_test(
    "Video with loading=lazy and autoplay does not play while not visible, then plays once scrolled into view"
  );
  let scrolledIntoView = false;

  t.step_timeout(() => {
    // Verify video has not played before scrolling
    t.step(() => {
      assert_true(video.paused, "Video should be paused when not visible in viewport");
    });

    // set timeout to 10000ms to fail if play event never fires
    const playTimeout = t.step_timeout(() => {
      assert_unreached("Play event never fired");
    }, 10000);

    // bind to play event to assert true that play has begun after scroll has been called
    video.addEventListener("play", () => {
      clearTimeout(playTimeout);
      t.step(() => {
        assert_true(scrolledIntoView, "Video should fire play event after scrolling into view");
      });
      t.done();
    });

    // scroll into view and set scrolledIntoView to true
    setTimeout(() => {
      scrolledIntoView = true;
      video.scrollIntoView();
    }, 1000);
  }, 2000);


</script>
</body>
</html>
